requirement.txt
Flask==2.2.5
mysql-connector-python==8.0.33
requests==2.31.0
Werkzeug==2.2.3